The windows implementation of the process wrapper now properly finds and shuts down any Windows Recovery Environment dialogs that are keeping H3D alive after crashes or freezes. psutil added as a requirement for running RunTests, UnitTest Installation.txt updated accordingly.

+++ trunk/H3DAPI/Util/UnitTests/ProcessWrapper.py
@@ -1,6 +1,7 @@
 ﻿""" Contains the Process interface definition as well as implementations of it for Unix and Windows """
 
-import os, sys
+import os
+import sys
 import subprocess
 import time
 import platform
@@ -19,41 +20,42 @@ if platform.system() == 'Windows':
   import win32gui
   import win32con
   import win32process
+  import psutil
 
 
-class Process ( object ):
+class Process(object):
   """ Interface for manipulating a process. """
 
-  def launch ( self, command, cwd ):
+  def launch(self, command, cwd):
     pass
 
-  def testLaunch ( self, command, cwd, startup_time, shutdown_time, startup_time_multiplier, early_shutdown_file ):
+  def testLaunch(self, command, cwd, startup_time, shutdown_time, startup_time_multiplier, early_shutdown_file):
     """ Makes sure early_shutdown_file doesn't exist, calls launch() and then waits for the specified startup and shutdown time. """
     pass
     
-  def isRunning ( self ):
+  def isRunning(self):
     """ Checks if the process associated with this object is running. """
     pass
     
-  def sendKey ( self, key ):
+  def sendKey(self, key):
     pass
     
-  def kill ( self ):
+  def kill(self):
     pass
     
-  def getStdOut ( self ):
+  def getStdOut(self):
     pass
     
-  def getStdErr ( self ):
+  def getStdErr(self):
     pass
     
 
 
-class ProcessWin32 ( Process ):
+class ProcessWin32(Process):
 
-  def launch ( self, command, cwd ):
+  def launch(self, command, cwd):
   
-    def enqueue_output ( stdout, stderr, std_out_q, std_err_q ):
+    def enqueue_output(stdout, stderr, std_out_q, std_err_q):
       for line in iter(stdout.readline, b''):
         std_out_q.put(line)
       stdout.close()
@@ -61,72 +63,71 @@ class ProcessWin32 ( Process ):
         std_err_q.put(line)
       stderr.close()
   
-    self.process_name= command[0]
+    self.process_name = command[0]
   
     shell = win32com.client.Dispatch("WScript.Shell")
 
     import ctypes
     ctypes.windll.kernel32.SetErrorMode(0)
 
-    self.process= subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd)
+    self.process = subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd)
     
     self.std_out_q = Queue()
     self.std_err_q = Queue()
     self.monitor_output_thread = Thread(target= enqueue_output, args=(self.process.stdout, self.process.stderr, self.std_out_q, self.std_err_q))
     self.monitor_output_thread.start()
     
-  def testLaunch ( self, command, cwd, startup_time, shutdown_time, startup_time_multiplier, early_shutdown_file ):
+  def testLaunch(self, command, cwd, startup_time, shutdown_time, startup_time_multiplier, early_shutdown_file):
   
     import ctypes
     SEM_NOGPFAULTERRORBOX = 0x0002 # From MSDN
     ctypes.windll.kernel32.SetErrorMode(SEM_NOGPFAULTERRORBOX)
     subprocess_flags = 0x8000000 #win32con.CREATE_NO_WINDOW?
     
-    self.process_name= command[0]
+    self.process_name = command[0]
     
-    if os.path.isfile( early_shutdown_file ):
-      os.remove( early_shutdown_file )
-    self.process= subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd, creationflags= subprocess_flags )
-    for i in range( 0, startup_time_multiplier ):
+    if os.path.isfile(early_shutdown_file):
+      os.remove(early_shutdown_file)
+    self.process = subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd, creationflags= subprocess_flags)
+    for i in range(0, startup_time_multiplier):
       time.sleep(startup_time)
-      if os.path.isfile( early_shutdown_file ) or not self.isRunning():
+      if os.path.isfile(early_shutdown_file) or not self.isRunning():
         break
-    success= self.isRunning()
+    success = self.isRunning()
     if success:
       self.kill()
     return success
     
-  def isRunning ( self ):
+  def isRunning(self):
     # for Windows XP use this
-    query_process = "tasklist /NH /FI \"IMAGENAME EQ "+ self.process_name +"\""
+    query_process = "tasklist /NH /FI \"IMAGENAME EQ " + self.process_name + "\""
     
     # for Windows Vista use this
     #query_process = 'query process'
     
     #for Linux use this
-    #query_process = 'ps -Af | grep ' + self.process_name + 
+    #query_process = 'ps -Af | grep ' + self.process_name +
     
-    p = subprocess.Popen( query_process, stdout=subprocess.PIPE, shell=True )
+    p = subprocess.Popen(query_process, stdout=subprocess.PIPE, shell=True)
     output = p.communicate()
     
     return output[0].find(self.process_name) >= 0
 
-  def sendKey ( self, key ):
+  def sendKey(self, key):
     if key == "{ESC}":
       # This will allow us to send ESC without the window being in focus
       target_pid = self.process.pid
       h3d_handle = 0
       
-      def win32_enum_handles_callback (handle):
-        _, found_pid = win32process.GetWindowThreadProcessId (handle)
-        if found_pid==target_pid:
+      def win32_enum_handles_callback(handle, dummy):
+        _, found_pid = win32process.GetWindowThreadProcessId(handle)
+        if found_pid == target_pid:
           h3d_handle = handle
         return True
       
       try:
-        win32gui.EnumWindows(win32_enum_handles_callback)
-        win32api.PostMessage(
-            h3d_handle, 
+        win32gui.EnumWindows(win32_enum_handles_callback, h3d_handle)
+        win32api.PostMessage(h3d_handle, 
             win32con.WM_KEYDOWN, 
             win32con.VK_ESCAPE, 
             0)
@@ -134,25 +135,38 @@ class ProcessWin32 ( Process ):
         print e
     else:
       shell = win32com.client.Dispatch("WScript.Shell")
-      shell.AppActivate ( self.process_name )
-      time.sleep ( 0.1 )
+      shell.AppActivate(self.process_name)
+      time.sleep(0.1)
       shell.SendKeys(key)
 
 
     
-  def kill ( self ):
-    # for Windows XP
-    kill_process = "taskkill /F /im "+ self.process_name
-    os.system( kill_process )
-    
-    # If the test crashed, close the crash dialog box too
-    shell = win32com.client.Dispatch("WScript.Shell")
-    shell.AppActivate ( self.process_name )
-    shell.SendKeys("{Esc}")
+  def kill(self):
+
+    # If the process crashed then windows might have started WerFault.exe that
+    # might be preventing the process from properly shutting down.
+    # We can solve this by looking for any WerFault process that was launched
+    # with the process id of our process and shut that down.
+    # Then if H3D is still running it means werfault wasn't keeping it alive
+    # and we can go ahead and do the taskkill.
+ 
+    for proc in psutil.process_iter():
+      if "WerFault.exe".lower() in proc.name().lower():
+        cmdparams = proc.cmdline()
+        for i in range(0, len(cmdparams)):
+          if (cmdparams[i] == "-p") and (i < len(cmdparams)-1) and (cmdparams[i+1] == str(self.process.pid)):
+            #This is the WerFault.exe that is forcefully keeping our process alive!
+            kill_process = "taskkill /F /PID " + str(proc.pid)
+            os.system(kill_process)
+        
+    if self.process.poll() == None:
+      kill_process = "taskkill /F /im " + self.process_name
+      os.system(kill_process)
 
 
-  def getStdOut ( self ):
-    output= ""
+
+  def getStdOut(self):
+    output = ""
     try:
       while not self.std_out_q.empty():
         output+= self.std_out_q.get(timeout=.1)
@@ -160,8 +174,8 @@ class ProcessWin32 ( Process ):
       pass
     return output
     
-  def getStdErr ( self ):
-    output= ""
+  def getStdErr(self):
+    output = ""
     try:
       while not self.std_err_q.empty():
         output+= self.std_err_q.get(timeout=.1)
@@ -171,10 +185,10 @@ class ProcessWin32 ( Process ):
 
 
 
-class ProcessUnix ( Process ):
-  def launch ( self, command, cwd ):
+class ProcessUnix(Process):
+  def launch(self, command, cwd):
       
-    def enqueue_output ( stdout, stderr, std_out_q, std_err_q ):
+    def enqueue_output(stdout, stderr, std_out_q, std_err_q):
       for line in iter(stdout.readline, b''):
         std_out_q.put(line)
       stdout.close()
@@ -182,58 +196,58 @@ class ProcessUnix ( Process ):
         std_err_q.put(line)
       stderr.close()
   
-    self.process_name= command[0]
+    self.process_name = command[0]
 
-    self.process= subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd)
+    self.process = subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd)
     
     self.std_out_q = Queue()
     self.std_err_q = Queue()
     self.monitor_output_thread = Thread(target= enqueue_output, args=(self.process.stdout, self.process.stderr, self.std_out_q, self.std_err_q))
     self.monitor_output_thread.start()
     
-  def testLaunch ( self, command, cwd, startup_time, shutdown_time, startup_time_multiplier, early_shutdown_file ):
+  def testLaunch(self, command, cwd, startup_time, shutdown_time, startup_time_multiplier, early_shutdown_file):
   
-    self.process_name= command[0]
-    if os.path.isfile( early_shutdown_file ):
-      os.remove( early_shutdown_file )
-    self.process= subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd )
-    for i in range( 0, startup_time_multiplier ):
+    self.process_name = command[0]
+    if os.path.isfile(early_shutdown_file):
+      os.remove(early_shutdown_file)
+    self.process = subprocess.Popen(command, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd= cwd)
+    for i in range(0, startup_time_multiplier):
       time.sleep(startup_time)
-      if os.path.isfile( early_shutdown_file ) or not self.isRunning():
+      if os.path.isfile(early_shutdown_file) or not self.isRunning():
         break
-    success= self.isRunning()
+    success = self.isRunning()
     if success:
       self.kill()
       time.sleep(shutdown_time)
     return success
     
-  def isRunning ( self ):
+  def isRunning(self):
     #for Linux use this
     query_process = 'ps -Af | grep ' + self.process_name
     
-    p = subprocess.Popen( query_process, stdout=subprocess.PIPE, shell=True )
+    p = subprocess.Popen(query_process, stdout=subprocess.PIPE, shell=True)
     output = p.communicate()
     output0 = output[0]
     # remove ps and grep process lines.
-    output0 = output0.replace( 'ps -Af | grep ' + self.process_name, "" )
-    output0 = output0.replace( 'grep ' + self.process_name, "" )
-    return output0.find(self.process_name) >= 0 and not (output0.find( '[' + self.process_name + '] <defunct>') >= 0 )
+    output0 = output0.replace('ps -Af | grep ' + self.process_name, "")
+    output0 = output0.replace('grep ' + self.process_name, "")
+    return output0.find(self.process_name) >= 0 and not (output0.find('[' + self.process_name + '] <defunct>') >= 0)
     
-  def sendKey ( self, key ):
-    time.sleep ( 0.1 )
+  def sendKey(self, key):
+    time.sleep(0.1)
     if key == "{ESC}" or key == "{Esc}":
 #      device = uinput.Device([uinput.KEY_ESC])
 #      device.emit_click(uinput.KEY_ESC)
       key_process = 'xte "key Escape"'
-      p = subprocess.Popen( key_process, stdout=subprocess.PIPE, shell=True )
+      p = subprocess.Popen(key_process, stdout=subprocess.PIPE, shell=True)
     else:
       raise Exception("Key " + key + " not handled exception")
     
-  def kill ( self ):
+  def kill(self):
     self.process.kill()
     
-  def getStdOut ( self ):
-    output= ""
+  def getStdOut(self):
+    output = ""
     try:
       while not self.std_out_q.empty():
         output+= self.std_out_q.get(timeout=.1)
@@ -241,8 +255,8 @@ class ProcessUnix ( Process ):
       pass
     return output
     
-  def getStdErr ( self ):
-    output= ""
+  def getStdErr(self):
+    output = ""
     try:
       while not self.std_err_q.empty():
         output+= self.std_err_q.get(timeout=.1)
